pyproject.toml
[project]
name = "notion-sync"
version = "0.1.0"
description = "Sync Notion database to Hugo content"
authors = [
    {name = "zi", email = "z4none@gmail.com"}
]
dependencies = [
    "notion-client>=2.2.1",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
    "rich>=13.7.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["notion_sync"]

[tool.hatchling.build.targets.sdist]
packages = ["notion_sync"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
notion_sync = "notion_sync.main:cli"
